requirements.txt
black==22.3.0
click==8.1.3
lxml==4.8.0
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
python-docx==0.8.11
tomli==2.0.1
